CGM: Add trait tag to override world rules
Add special handling for the `CGM_OverrideWorldRules_AlwaysAllow` trait
tag so that CGM allows the trait to be selected even if it's disallowed
by world rules. Exclusive traits are still honoured.

The main purpose of this trait tag is to have traits that cannot appear
by default on planets, but that can be manually added through CGM. This
can be accomplished with the following trait tags:

  - CGM_OverrideWorldRules_AlwaysAllow
  - StartWorldOnly
  - NonStartWorld

Signed-off-by: Angel Pons <th3fanbus@gmail.com>