MSBuild incremental and automatic Publicise / Refasmer
This updates the `Publicise` and `GenerateRefAssemblies` tasks to run
automatically when their output files are missing or out-of-date
relative to their input files.

MSBuild will batch all the changes into a single operation, so one call
to the task will be made for all the inputs that need rebuilding, but
it'll be limited to things where the timestamp suggests it is newer.

This required lifting the ItemGroup containing their inputs out of the
Target element, because MSBuild depressingly doesn't expose the Target
Inputs and Outputs properties to the contained Tasks.  -___-

The tasks are also now conditional on the thing being built being
flagged with `IsMod`, which seems the correct behaviour to me.