Manage both Publicise and ILRepack nuget packages centrally
This is a very large *looking* change, but the functional difference is
small: in `Directory.Build.Props` the packages for the two
"infrastructure" tasks are managed centrally:

- Aze.Publicise.MSBuild.Task
- ILRepack.Lib.MSBuild.Task

The rest of the change is removing the `PackageReference Update=...`
elements that MSVC introduced to every project when an update to the
version was specified.

To preserve this it is necessary to manually update the version number
in `Directory.Build.Props` rather than letting the MSVC "Manage NuGet
packages for..." tool perform the task.