Always use "latest stable" version of Publicise / ILRepack tasks
Instead of specifying a version manually, the global package reference
is now for `Version="*"`.  That is the "floating version" matching
"The highest stable version."

https://learn.microsoft.com/en-us/nuget/concepts/dependency-resolution#floating-versions

This'll ensure that manual updates to the package versions for those
tasks are not required, at least until the day they ship an incompatible
update and need to be pinned back.

That seems a much less common situation than needing to update them over
time though.